requirements.txt
# Core dependencies
loguru>=0.7.3
python-dateutil>=2.8.0
opentelemetry-api>=1.37.0
opentelemetry-sdk>=1.37.0
opentelemetry-exporter-otlp>=1.37.0
tiktoken>=0.5.0

# Web framework integration
starlette>=0.20.0
fastapi>=0.118.0